Reverted to edit 4ede2159-c0d1-4385-bf0e-4f16e2e7b1f7: "Merge pull request #31 from catfandoglover/dashboard
Dashboard -- totally new so shouldn't break anything else from what I gathered"
